requirements.txt
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
cycler==0.12.1
decorator==5.2.1
easydict==1.13
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
huggingface-hub==0.30.1
idna==3.10
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lmdb==1.6.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
termcolor==3.0.1
threadpoolctl==3.6.0
timm==1.0.15
torch==2.3.1
torchvision==0.18.1
tqdm==4.67.1
traitlets==5.14.3
triton==2.3.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
yacs==0.1.8
